requirements.txt
numpy
pandas
scipy
scikit-learn
statsmodels
matplotlib
seaborn
plotly
networkx
pyyaml
tqdm
torch
torch-geometric
holidays
tweepy
requests
openai
semanticscholar>=0.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
thefuzz>=0.22.0
